ref(seer): Merge state context into sessions context
Fold seerExplorerStateContext into seerExplorerSessionContext so there
is a single provider that owns the query, reducer, and merge logic.
The context value is now {query, conversations}, letting consumers
destructure what they need. Removes a provider nesting layer from
the org layout.
